refactor(context): refactor hidden and filter
